Add 'gist' for git submodule top, 'gisd' for debug.
'gis' identifies the 'top' of a repository which involves submodules;
i.e., furthest parent directory w/ a .gitmodules file.

If $0 is 'gist' then the argument is executed in the 'top' repository.

If $0 is 'gisd' then sanity checks for submodules are skipped and the
commands are executed across all submodules and then the 'top' repo.

If $1 is 'branches' a summary of merge state is printed for all local
branches.